Remove serialization classes from Message
Split out the MessageJsonSerializer and MessageJsonDeserializer into two
standalone classes that include their static helper methods.
